#9045: Merge 1 and 2 cq tests back into one file after enabling mixing of 1/2 cqs in the same process
