+++ tests/scripts/run_performance.sh
@@ -17,9 +17,7 @@ run_perf_models_other() {
 
     env pytest -n auto models/demos/ttnn_falcon7b/tests -m $test_marker
 
-    # Separate calls since we can't mix switching between number of cqs
     env pytest -n auto models/demos/resnet/tests/test_perf_resnet.py -m $test_marker
-    env pytest -n auto models/demos/resnet/tests/test_perf_resnet_2cqs.py -m $test_marker
 
     env pytest -n auto tests/ttnn/integration_tests/whisper/test_performance.py -m $test_marker
