+++ tests/scripts/single_card/nightly/run_gs_only.sh
@@ -12,8 +12,6 @@ echo "Running model nightly tests for GS only"
 
 env pytest -n auto models/demos/resnet/tests/test_metal_resnet50_performant.py ; fail+=$?
 
-env pytest -n auto models/demos/resnet/tests/test_metal_resnet50_2cqs_performant.py ; fail+=$?
-
 if [[ $fail -ne 0 ]]; then
   exit 1
 fi
